Fix broken arrow dropdown when used with Preact
This commit provides a fix for an issue with this code:

handleMouseDownOnArrow (event) {
	// if the event was triggered by a mousedown and not the primary
	// button, or if the component is disabled, ignore it.
	if (this.props.disabled || (event.type === 'mousedown' && event.button !== 0)) {
		return;
	}
	// If the menu isn't open, let the event bubble to the main handleMouseDown
	if (!this.state.isOpen) {
		this.setState({
			isOpen: true,
		});
	}
	// prevent default event handlers
	event.stopPropagation();
	event.preventDefault();
	// close the menu
	if(this.state.isOpen){
		this.closeMenu();
	}
}
```

When trying to open the select and in React, the
this.setState({ isOpen: true}) call will change the state async, whereas
in Preact this happens immediately. This means that, in React, the
expression in the line if(this.state.isOpen){ will be false, but in
Preact it's true. So state.isOpen is flicked to true and then straight
back to false before render() is called. Meaning that in Preact this
code has no effect.